Enable parallel (OpenMP) ray-casting in SCM.
Note that Bullet ray-casting (in the version currently embedded in Chrono) does not appear to be thread-safe when intersecting a collision mesh.  As such, we are limited to parallelize only the top-level loop (over the moving patches).  Ideally, we would parallelize over all rays cast in each domain (corresponding to each moving patch).